refactor(metadata): Move DeleteDeviceResourceByName and DeleteDeviceCommandByName to correct files
move from deviceprofile.go to deviceresource.go and devicecommand.go

close #3970
